Move utils/pulse to metabase/pulse since it contains only domain-related code
