+++ enterprise/frontend/src/metabase-enterprise/sharing/components/MutableParametersSection.tsx
@@ -7,8 +7,8 @@ import { CollapseSection } from "metabase/common/components/CollapseSection";
 import CS from "metabase/css/core/index.css";
 import { ParametersList } from "metabase/parameters/components/ParametersList";
 import { getVisibleParameters } from "metabase/parameters/utils/ui";
+import { getPulseParameters } from "metabase/pulse";
 import type { DraftDashboardSubscription } from "metabase/redux/store";
-import { getPulseParameters } from "metabase/utils/pulse";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import { deriveFieldOperatorFromParameter } from "metabase-lib/v1/parameters/utils/operators";
 import {
