+++ frontend/src/metabase/account/notifications/components/NotificationCard/DashboardNotificationCard.tsx
@@ -6,7 +6,7 @@ import { formatCreatorMessage } from "metabase/account/notifications/components/
 import type { DashboardSubscriptionListItem } from "metabase/account/notifications/types";
 import { Link } from "metabase/common/components/Link";
 import { formatTitle } from "metabase/notifications/utils";
-import { canArchiveLegacyAlert, formatChannel } from "metabase/utils/pulse";
+import { canArchiveLegacyAlert, formatChannel } from "metabase/pulse";
 import * as Urls from "metabase/utils/urls";
 import type { Channel, User } from "metabase-types/api";
