+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/DashboardSubscriptionsSidebar.tsx
@@ -17,14 +17,10 @@ import {
   getEditingPulse,
   getPulseFormInput,
 } from "metabase/notifications/pulse/selectors";
+import { NEW_PULSE_TEMPLATE, cleanPulse, createChannel } from "metabase/pulse";
 import type { DraftDashboardSubscription, State } from "metabase/redux/store";
 import { getUser, getUserIsAdmin } from "metabase/selectors/user";
 import { UserApi } from "metabase/services";
-import {
-  NEW_PULSE_TEMPLATE,
-  cleanPulse,
-  createChannel,
-} from "metabase/utils/pulse";
 import { connect } from "metabase/utils/redux";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import type {
