+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/PulsesListSidebar.tsx
@@ -7,6 +7,7 @@ import { Label } from "metabase/common/components/type/Label";
 import { Subhead } from "metabase/common/components/type/Subhead";
 import CS from "metabase/css/core/index.css";
 import { formatDateValue } from "metabase/parameters/utils/date-formatting";
+import { getActivePulseParameters } from "metabase/pulse";
 import type { State } from "metabase/redux/store";
 import { Button, Icon, Tooltip } from "metabase/ui";
 import {
@@ -14,7 +15,6 @@ import {
   formatDateTimeWithUnit,
   formatTimeWithUnit,
 } from "metabase/utils/formatting";
-import { getActivePulseParameters } from "metabase/utils/pulse";
 import { connect } from "metabase/utils/redux";
 import { formatFrame } from "metabase/utils/time-dayjs";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
