+++ frontend/src/metabase/notifications/modals/CreateOrEditQuestionAlertModal/CreateOrEditQuestionAlertModal.tsx
@@ -17,6 +17,10 @@ import {
   getAlertTriggerOptions,
   getDefaultQuestionAlertRequest,
 } from "metabase/notifications/utils";
+import {
+  getHasConfiguredAnyChannel,
+  getHasConfiguredEmailOrSlackChannel,
+} from "metabase/pulse";
 import { updateUrl } from "metabase/query_builder/actions/url";
 import {
   getQuestion,
@@ -36,10 +40,6 @@ import {
   rem,
 } from "metabase/ui";
 import { getResponseErrorMessage } from "metabase/utils/errors";
-import {
-  getHasConfiguredAnyChannel,
-  getHasConfiguredEmailOrSlackChannel,
-} from "metabase/utils/pulse";
 import { useDispatch, useSelector } from "metabase/utils/redux";
 import type Question from "metabase-lib/v1/Question";
 import type {
