+++ frontend/src/metabase/pulse/index.ts
@@ -1,0 +1,1 @@
+export * from "./pulse";
